Rationalise class names
Classes should never have capitals! Unless you are using camelCase, but that is not standard for classes or ids generally.

camelCase for JS vars
kebab-case for classes and ids (so called cos it’s on a skewer -like-a-tasty-kebab-)

this means if I see jonnysBigJob I know that this is a JS var, but jonnys-big-job will definitely be a class or ID.