Let’s automate!
First, let’s remove ‘_choice’ from the IDs in index.html. It just complicates things unneccessarily.

We create an array of our dom identifiers (`options`), and an empty array to hold the element references (`buttons`).

We call them buttons because an interactive element we can click on is a button. We can make them actual buttons in the DOM later.

Then we loop through the dom identifiers, assigning the dom element consttructed from the identifier, to the `buttons` array item, with a key that is the identifier too.

Later we replace all the individual onclick functions with a loop that does the same.
We take the id of the event target and use that to set myChoice.
`var myChoice = event.target.id;`