Switch 3: To switch, or not to switch?
So why not use switch for the `computerChoice` logic block?

Switch is designed to check equality does `var` equal `’value’`?
```switch(var){
    case(‘value’):
        // do the thing
        break;```

It is NOT good for non equality comparison:
http://stackoverflow.com/questions/6665997/switch-statement-for-greater-than-less-than

HOO! THAT BE SLOOOOOOW!

Ok, so can we do anything?
Not a lot, but still something.

The `else if(0.21 <= computerChoice <= 0.40){` checks are kind a cute, but why are we doing them. It’s an if else, anything less than 0.21 has been dealt with!

If we did do that, why `0.21 <=` ? Why not `2.0 <`

One other problem:
```    if (computerChoice < 0.20){
        // blah
    }
    else if(0.21 <= computerChoice <= 0.40){
        // blah 2
    }```
What if it’s === 0.2? Whoops! Small hole there.
And the same for 0.4 and 0.6!

Never mind, we’ll clear that up.

None of it belongs in `compare()` anyway, so let’s break that out too.